Only merge in tango changes in this commit. bm26a-config changes to
follow. Ignore b21-config changes

+++ plugins/org.gda.tango/src/gda/device/lima/LimaCCD.java
@@ -217,9 +217,9 @@ public interface LimaCCD extends Base {
 
 	void setSavingOverwritePolicy(SavingOverwritePolicy savingOverwritePolicy) throws DevFailed;
 
-	long getSavingFramePerFile() throws DevFailed;
+	int getSavingFramePerFile() throws DevFailed;
 
-	void setSavingFramePerFile(long savingFramePerFile) throws DevFailed;
+	void setSavingFramePerFile(int savingFramePerFile) throws DevFailed;
 
 	String[] getSavingCommonHeader() throws DevFailed;
